requirements.txt
boto3==1.38.24
botocore==1.38.24
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.41
alembic==1.16.1
pydantic>=2.7.0
python-dotenv==1.0.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
psycopg2-binary==2.9.9
pydantic-settings==2.9.1
mypy_boto3_cognito_idp==1.38.16
uuid==1.30
